Fix a bug where features are too narrow (RT341295)
This reverts a small part of a change that was made in commit 90ea866 that caused a bug where features that are the minimum width are too narrow, so you can't see the fill colour. The problem is to do with the offset that is applied to draw the outline. I'm not sure why the change was made as it was part of a large commit, but undoing this small part seems to revert the behaviour to how it was previously.