Fix a bug with variations not being passed to blixem
Looks like this was broken in the GFF3 changes. The problem was that the SO term was being checked against the SO ID, not the term name. There was also a problem with some variation types not being included in the blixem output. The list is hard-coded in the blixem code at the moment but in the long term we should use the GFF dumper code instead.