Improve handling of relative paths and non-writeable locations for config files - NEW FEATURE 'Improve config file handling'
Previously the conf_dir had to be writable if given because the log was created there. Now the log is created in ~/.ZMap if the conf_dir is not writable.

The path for the styles file can now be relative to the current directory as well as relative to the conf_dir.