Reorganise edit-styles dialog code - WIP
Removed complicated rules for whether to create a new child style or not. It now allows the user to enter the style name to edit. This defaults to the current style if it's the same as the featureset name, or to creating a new child style with the featureset name if the style name is different.