Remove warning when source is empty - BUG FIX
Previously there was a misleading error about the source with non-friendly text about context being empty. Users probably don't want to be bothered when it's just that there were no features, so I've removed it. Alternativly we could issue a friendlier message now I've returned an error code and we know what the error is.