Better blixem BAM configuration - NEW FEATURE 'BAM config'
When you call blixem on a bam column, the bam_get command to call is now passed in a 'command' tag in the gff file, rather than passing the dataType=short-read tag and relying on the caller having set up a short-read data type in the blixem config file. I've also removed the special right-click menu items that were bam specific. The normal menu items and shortcuts can now be used for bam columns.

Following on from merge of most of the work in feature/bam. This last commit fixes a problem where the bam options weren't being called when using the keyboard shortcuts.