Allow GFF files without a sequence-region comment - NEW FEATURE 'GFF sequence-region'
Previously, zmap required GFF files to have a sequence-region comment. However, this is optional in the GFF3 spec so we now support files without this comment.

If a sequence region is specified in the config/command-line then zmap will load any features from the file that overlap that region.

If no sequence region is specified elsewhere, then zmap will prompt the user to enter the sequence details.