Re-add hack to add source data for bam sources - WIP
This is required for sources which use bam_get because there is no other way to set the is_seq flag to indicate that it's a bam source. This could be done in a better way but preserving the old behaviour for now as bam_get will be removed eventually.