Remove a thread's connection straight away after killing it - WIP
Previously thread_has_died was not being set so the connection was being removed on a subsequent iteration through checkStateConnections, causing confusing error messages in the log because there is no error message the second time around. I'm not sure if there was a reason to wait but it seems sensible to remove the connection after killing the thread.